Add JEOS_EULA_ALREADY_AGREED option
We could need to create images that leverage jeos-firstboot to take care
of locale / keyboard or network configuration, and where the image is
already part of a different product.  For example, a PXE boot image
(installed via RPM), that will be used as a component to provisioning
new images in a network.

This use case will require no user interaction during the boot process,
and currently there is a dialog that forces the user to accept the EULA
license.

This patch add a new option, JEOS_EULA_ALREADY_AGREED, that if set to a
nonempty value, will skip the dialog showed for EULA acceptance. As
commented in the example configuration file, this option is expected to
be used only when the image is part of a different product that contains
an already accepted license.

Fix #63