Resolve cycle w.r.t. wicked
cloud-init.target include cloud-init.service which needs to run after wicked.
But this service wants to run prior to wicked. However, cloud-init knows
what it shold be doing after the cloud-init-local phase and that service
runs prior to wicked. As such the requirements from a systemd timing
perspective are consistent.